# docs/desktop-auth-openapi.yaml
openapi: 3.1.0
info:
  title: Desktop Authentication API
  # version tracks common.Version (common/constants.go), which defaults to
  # "v0.0.0" and is overridden at build time. Keep the number in sync when the
  # auth contract changes.
  version: "0.0.0"
  description: |
    Authentication entry points for the desktop (Electron) client of this AI
    API gateway. These endpoints issue short-lived access tokens and 7-day
    desktop refresh sessions, fully separated from the web browser session
    routes under `/api/user/auth`.

    The four desktop auth endpoints (`/api/desktop/auth/*`) share a single
    JSON envelope:

    ```json
    {
      "success": true,
      "code": "OK",
      "message": "",
      "request_id": "01J...",
      "server_time": 1730000000,
      "data": {}
    }
    ```

    The two bootstrap endpoints (`GET /api/status` and
    `GET /api/user/login/encryption-key`) are pre-existing web routes reused by
    the desktop client; they return the legacy web envelope
    `{ "success": true, "message": "", "data": {} }` with no `code`,
    `request_id`, or `server_time`.

    Every desktop auth response carries `Cache-Control: no-store`. Never cache
    or log tokens. Requests to all four `/api/desktop/auth/*` endpoints are
    rate limited by the `DesktopAuthRateLimit` middleware; exceeding the budget
    yields `429` `AUTH_RATE_LIMITED` (with a `Retry-After` header).

    Password submission: either send the raw `password`, or — when password
...
servers:
...
paths:
...
  /api/desktop/auth/login:
    post:
      tags: [Desktop Auth]
      operationId: desktopLogin
      summary: Exchange username/password for a desktop session
      description: |
        Authenticates a desktop client. Requires the device fingerprint
        fields. If the account requires two-factor authentication, the response
        is `200` with `code: AUTH_VERIFICATION_REQUIRED` and a challenge
        (`data.flow_token`); finish it with `/api/desktop/auth/verify`.
        Otherwise a full token bundle is returned. The two `data` shapes are
        discriminated by `data.require_verification`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesktopLoginRequest"
            example:
              username: "alice@example.com"
              password_encrypted: "v2.BASE64WRAPPEDKEY.BASE64NONCE.BASE64CIPHERTEXT"
              encryption_key_id: "0123456789abcdef0123456789abcdef"
              device_id: "ae4c1f2b-9c3a-4d7e-8b2a-1f0e2d3c4b5a"
              device_name: "Alice's MacBook Pro"
              platform: "darwin"
              arch: "arm64"
              client_version: "1.2.0"
              captcha_token: "0.ChM...turnstile-response..."
      responses:
        "200":
          description: |
            Login succeeded. `data` is either a token bundle
            (`code: OK`) or a verification challenge
            (`code: AUTH_VERIFICATION_REQUIRED`).
          headers:
            Cache-Control:
              schema: { type: string }
              example: "no-store"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: "#/components/schemas/DesktopAuthSuccessData"
                          - $ref: "#/components/schemas/DesktopVerificationChallengeData"
                        discriminator:
                          propertyName: require_verification
                          mapping:
                            "true": "#/components/schemas/DesktopVerificationChallengeData"
                            "false": "#/components/schemas/DesktopAuthSuccessData"
              examples:
                login_success:
                  $ref: "#/components/examples/login_success"
                login_totp_challenge:
                  $ref: "#/components/examples/login_totp_challenge"
        "400":
          description: |
            Bad request. Error codes: `INVALID_ARGUMENT` (malformed JSON,
            missing/blank required fields, invalid device metadata, or
            mutually exclusive password fields), `AUTH_CAPTCHA_INVALID`
            (Turnstile token rejected), `AUTH_ENCRYPTION_PAYLOAD_INVALID`
            (bad v2 envelope, encryption enabled but fields incomplete, or
            encryption disabled but encrypted fields sent).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "INVALID_ARGUMENT"
                message: "Invalid request"
                request_id: "01J8XQ2K7VZ9W3Y4T5S6A7B8C9"
                server_time: 1730000000
                data: null
        "401":
          description: Invalid username or password (`AUTH_INVALID_CREDENTIALS`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_INVALID_CREDENTIALS"
                message: "Invalid username or password"
                request_id: "01J8XQ2K7VZ9W3Y4T5S6A7B8C9"
                server_time: 1730000000
                data: null
        "403":
          description: |
            `AUTH_PASSWORD_LOGIN_DISABLED` (password login turned off on the
            server), `AUTH_CAPTCHA_REQUIRED` (Turnstile enabled but
            `captcha_token` omitted), or `AUTH_VERIFICATION_UNSUPPORTED` (no
            usable second factor available).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_CAPTCHA_REQUIRED"
                message: "Captcha is required"
                request_id: "01J8XQ2K7VZ9W3Y4T5S6A7B8C9"
                server_time: 1730000000
                data: null
        "409":
          description: |
            `AUTH_ENCRYPTION_KEY_STALE` (client used an outdated encryption
            public key; re-fetch `GET /api/user/login/encryption-key` and
            retry), or `AUTH_SESSION_LIMIT` (too many active desktop sessions).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_ENCRYPTION_KEY_STALE"
                message: "Password encryption key is stale"
                request_id: "01J8XQ2K7VZ9W3Y4T5S6A7B8C9"
                server_time: 1730000000
                data: null
        "429":
          description: |
            Rate limited (`AUTH_RATE_LIMITED` from DesktopAuthRateLimit) or
            `AUTH_SESSION_ISSUANCE_LIMIT` (too many sessions created in the
            issuance window). Back off and retry after the `Retry-After`
            seconds.
          headers:
            Retry-After:
              schema: { type: integer, format: int32 }
              description: Minimum seconds to wait before retrying.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              examples:
                rate_limited_429:
                  $ref: "#/components/examples/rate_limited_429"
        "500":
          description: Internal error (`AUTH_INTERNAL_ERROR`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              examples:
                internal_error_5xx:
                  $ref: "#/components/examples/internal_error_5xx"
        "503":
          description: |
            Turnstile verification service unreachable (`AUTH_CAPTCHA_UNAVAILABLE`,
            emitted by this handler). Separately, the deployment/edge gateway MAY
            return 503 with `code: SERVICE_UNAVAILABLE` when the service is down;
            that code is NOT emitted by any current handler and is a pending
            deployment-layer contract (see `x-status`).
          x-status: pending-gateway-layer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_CAPTCHA_UNAVAILABLE"
                message: "Captcha service is unavailable"
                request_id: "01J8XQ2K7VZ9W3Y4T5S6A7B8C9"
                server_time: 1730000000
                data: null
  /api/desktop/auth/verify:
    post:
      tags: [Desktop Auth]
      operationId: desktopVerify
      summary: Complete a TOTP login challenge
      description: |
        Submits the TOTP code for a `flow_token` returned by `login`. On
        success, exchanges the challenge for a full token bundle.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesktopVerifyRequest"
            example:
              flow_token: "01J8XQ3M8WZ9X4Y5T6S7A8B9C0D"
              method: "totp"
              code: "123456"
      responses:
        "200":
          description: Challenge completed; a fresh token bundle is issued.
          headers:
            Cache-Control:
              schema: { type: string }
              example: "no-store"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DesktopAuthSuccessData"
              examples:
                totp_verify_success:
                  $ref: "#/components/examples/totp_verify_success"
        "400":
          description: "Malformed request (`INVALID_ARGUMENT`): missing `flow_token`, `code`, or `method` is not `totp`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "INVALID_ARGUMENT"
                message: "Invalid request"
                request_id: "01J8XQ4N9XA0Y5Z6U7V8B9C0D1E2"
                server_time: 1730000060
                data: null
        "401":
          description: |
            `AUTH_FLOW_EXPIRED` (challenge expired, consumed, or unknown) or
            `AUTH_VERIFICATION_FAILED` (wrong TOTP code, locked account, or
            2FA not enabled).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              examples:
                totp_verify_failed:
                  $ref: "#/components/examples/totp_verify_failed"
        "403":
          description: "`AUTH_VERIFICATION_UNSUPPORTED` (the requested second factor is not available)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_VERIFICATION_UNSUPPORTED"
                message: "Verification method is unsupported"
                request_id: "01J8XQ4N9XA0Y5Z6U7V8B9C0D1E2"
                server_time: 1730000060
                data: null
        "409":
          description: Session state conflict (`AUTH_SESSION_MISMATCH` / `AUTH_REFRESH_RACE`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_SESSION_MISMATCH"
                message: "Session credentials do not match"
                request_id: "01J8XQ4N9XA0Y5Z6U7V8B9C0D1E2"
                server_time: 1730000060
                data: null
        "429":
          description: Rate limited (`AUTH_RATE_LIMITED`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_RATE_LIMITED"
                message: "Too Many Requests"
                request_id: "01J8XQ4N9XA0Y5Z6U7V8B9C0D1E2"
                server_time: 1730000060
                data: null
        "500":
          description: Internal error (`AUTH_INTERNAL_ERROR`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_INTERNAL_ERROR"
                message: "Internal Server Error"
                request_id: "01J8XQ4N9XA0Y5Z6U7V8B9C0D1E2"
                server_time: 1730000060
                data: null
        "503":
          description: |
            Optional deployment/gateway-layer response only. The edge gateway or
            circuit breaker MAY return 503 with `code: SERVICE_UNAVAILABLE` when
            the service is down. No current verify handler emits this code; it is
            a pending deployment-layer contract and must not be asserted as backend
            behavior.
          x-status: pending-gateway-layer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "SERVICE_UNAVAILABLE"
                message: "Service temporarily unavailable"
                request_id: ""
                server_time: 0
                data: null
  /api/desktop/auth/refresh:
    post:
      tags: [Desktop Auth]
      operationId: desktopRefresh
      summary: Rotate a desktop refresh token
      description: |
        Rotates the refresh token and issues a new access token. The refresh
        token is single-use: a successful response invalidates the presented
        token and returns a new one. A short (~30s) tolerance window replays
        the immediately previous token idempotently, returning the same bundle;
        any replay outside that window revokes the whole session
        (`AUTH_SESSION_REVOKED`).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesktopRefreshRequest"
            example:
              refresh_token: "8c4d0b2e-3f5a-6b71-8293-0b1c2d3e4f5a.ANOTHER...SECRET"
              sid: "8c4d0b2e-3f5a-6b71-8293-0b1c2d3e4f5a"
      responses:
        "200":
          description: Rotated; a new access/refresh bundle is returned (or the same bundle on in-window replay).
          headers:
            Cache-Control:
              schema: { type: string }
              example: "no-store"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DesktopAuthSuccessData"
              examples:
                refresh_rotation:
                  $ref: "#/components/examples/refresh_rotation"
                refresh_30s_retry:
                  $ref: "#/components/examples/refresh_30s_retry"
        "400":
          description: Missing `refresh_token` or `sid` (`INVALID_ARGUMENT`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "INVALID_ARGUMENT"
                message: "Invalid request"
                request_id: "01J8XQ5P0YB1Z6A8W9X0C1D2E3F4"
                server_time: 1730003600
                data: null
        "401":
          description: |
            `AUTH_UNAUTHORIZED` (malformed/unknown refresh token),
            `AUTH_TOKEN_EXPIRED`, `AUTH_SESSION_EXPIRED`, or
            `AUTH_SESSION_REVOKED` (session expired, revoked, or rotated after a
            credential change / out-of-window replay). The client must re-login.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              examples:
                refresh_out_of_window_replay:
                  $ref: "#/components/examples/refresh_out_of_window_replay"
        "409":
          description: "`AUTH_SESSION_MISMATCH` (`sid` does not match the token) or `AUTH_REFRESH_RACE` (concurrent rotation)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_REFRESH_RACE"
                message: "Conflict"
                request_id: "01J8XQ5P0YB1Z6A8W9X0C1D2E3F4"
                server_time: 1730003600
                data: null
        "429":
          description: Rate limited (`AUTH_RATE_LIMITED`) or issuance window exhausted (`AUTH_SESSION_ISSUANCE_LIMIT`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_RATE_LIMITED"
                message: "Too Many Requests"
                request_id: "01J8XQ5P0YB1Z6A8W9X0C1D2E3F4"
                server_time: 1730003600
                data: null
        "500":
          description: Internal error (`AUTH_INTERNAL_ERROR`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_INTERNAL_ERROR"
                message: "Internal Server Error"
                request_id: "01J8XQ5P0YB1Z6A8W9X0C1D2E3F4"
                server_time: 1730003600
                data: null
        "503":
          description: |
            Optional deployment/gateway-layer response only. The edge gateway or
            circuit breaker MAY return 503 with `code: SERVICE_UNAVAILABLE` when
            the service is down. No current refresh handler emits this code; it is
            a pending deployment-layer contract and must not be asserted as backend
            behavior.
          x-status: pending-gateway-layer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "SERVICE_UNAVAILABLE"
                message: "Service temporarily unavailable"
                request_id: ""
                server_time: 0
                data: null
  /api/desktop/auth/logout:
    post:
      tags: [Desktop Auth]
      operationId: desktopLogout
      summary: Revoke a desktop session
      description: |
        Revokes the session identified by the refresh token. Optionally send
        the current access token as `Authorization: Bearer <access_token>`; if
        present and it does not match the refresh token's session, the request
        fails with `409 AUTH_SESSION_MISMATCH` (anti-CSRF check). Logout is
        idempotent: revoking an already-revoked or unknown session still
        returns `200 {logged_out: true}`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesktopLogoutRequest"
            example:
              refresh_token: "9d5e1c3f-4a6b-7c82-9304-1c2d3e4f5a6b.ROTATED...SECRET"
              sid: "8c4d0b2e-3f5a-6b71-8293-0b1c2d3e4f5a"
      responses:
        "200":
          description: Session revoked (or already gone; logout is idempotent for unknown/revoked sessions).
          headers:
            Cache-Control:
              schema: { type: string }
              example: "no-store"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DesktopLogoutData"
              examples:
                logout_success:
                  $ref: "#/components/examples/logout_success"
                logout_duplicate:
                  $ref: "#/components/examples/logout_duplicate"
        "400":
          description: Missing `refresh_token` or `sid` (`INVALID_ARGUMENT`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "INVALID_ARGUMENT"
                message: "Invalid request"
                request_id: "01J8XQ6Q2ZC2A7B0Y1Z2D3E4F5A6"
                server_time: 1730004000
                data: null
        "401":
          description: Refresh token not recognized for a desktop session (`AUTH_UNAUTHORIZED`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              examples:
                logout_random_token:
                  $ref: "#/components/examples/logout_random_token"
        "409":
          description: "`AUTH_SESSION_MISMATCH` (presented access token belongs to a different session, or `sid` does not match the refresh token)."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_SESSION_MISMATCH"
                message: "Session credentials do not match"
                request_id: "01J8XQ6Q2ZC2A7B0Y1Z2D3E4F5A6"
                server_time: 1730004000
                data: null
        "429":
          description: Rate limited (`AUTH_RATE_LIMITED`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_RATE_LIMITED"
                message: "Too Many Requests"
                request_id: "01J8XQ6Q2ZC2A7B0Y1Z2D3E4F5A6"
                server_time: 1730004000
                data: null
        "500":
          description: Internal error (`AUTH_INTERNAL_ERROR`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "AUTH_INTERNAL_ERROR"
                message: "Internal Server Error"
                request_id: "01J8XQ6Q2ZC2A7B0Y1Z2D3E4F5A6"
                server_time: 1730004000
                data: null
        "503":
          description: |
            Optional deployment/gateway-layer response only. The edge gateway or
            circuit breaker MAY return 503 with `code: SERVICE_UNAVAILABLE` when
            the service is down. No current logout handler emits this code; it is
            a pending deployment-layer contract and must not be asserted as backend
            behavior.
          x-status: pending-gateway-layer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
              example:
                success: false
                code: "SERVICE_UNAVAILABLE"
                message: "Service temporarily unavailable"
                request_id: ""
                server_time: 0
                data: null
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Optional access token for logout's refresh-token/access-token binding check.
  schemas:
    Envelope:
      type: object
      description: Shared response envelope returned by every /api/desktop/auth/* endpoint.
      required: [success, code, message, request_id, server_time]
      properties:
        success:
          type: boolean
          description: Whether the request succeeded.
        code:
          type: string
          description: Machine-readable status code; "OK" on success.
        message:
          type: string
          description: Human-readable message; empty on success.
        request_id:
          type: string
          description: Correlation ID for this request.
        server_time:
          type: integer
          format: int64
          description: Server Unix timestamp when the response was written.
        data:
          description: Endpoint payload; null on error.
...
    DesktopLogoutData:
      type: object
      properties:
        logged_out:
          type: boolean
          example: true
  examples:
...
    login_totp_challenge:
      summary: TOTP challenge required
      value:
        success: true
        code: "AUTH_VERIFICATION_REQUIRED"
        message: "Verification required"
        request_id: "01J8XQ2K7VZ9W3Y4T5S6A7B8C9"
        server_time: 1730000000
        data:
          require_verification: true
          flow_token: "01J8XQ3M8WZ9X4Y5T6S7A8B9C0D"
          expires_at: 1730000300
          methods: ["totp"]
...